twitterwebservice/requirements.txt
django~=2.2
aiodns==1.1.1
aiohttp==3.2.1
astroid==1.6.3
async-timeout==3.0.0
attrs==18.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
cassandra-driver==3.14.0
cchardet==2.1.1
certifi==2018.4.16
chardet==3.0.4
clearbit==0.1.7
click==6.7
click-plugins==1.0.3
colorama==0.3.9
django-cors-headers==2.2.0
elasticsearch==6.2.0
facebook-sdk==2.0.0
flextls==0.3
FullContact.py==0.0.6
html2text==2018.1.9
idna==2.6
#idna-ssl==1.0.1
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.1
mccabe==0.6.1
multidict==4.3.1
nltk==3.3
numpy==1.14.3
oauthlib==2.0.7
pandas==0.22.0
Pillow==5.1.0
pycares==2.3.0
pylint==1.8.4
pyodbc==4.0.23
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
shodan==1.7.7
six==1.11.0
sslscan==0.5.1
steganography==0.1.1
textblob==0.15.1
tweepy==3.6.0
urllib3==1.22
wrapt==1.10.11
XlsxWriter==1.0.5
yarl==1.2.4
langdetect==1.0.8
